+++ trunk/awesome-tim/awesome-tim/Plataforma.cpp
@@ -1,6 +1,6 @@
 #include "Plataforma.h"
 
-Plataforma::Plataforma(double largo,double posX,double posY,double angulo):Figura(ID_PLATAFORMA,new Cuadrado(ANCHO_PLATAFORMA*largo,ALTO_PLATAFORMA,posX,posY,angulo),false){
+Plataforma::Plataforma(int largo,double posX,double posY,double angulo):Figura(ID_PLATAFORMA,new Cuadrado(ANCHO_PLATAFORMA*largo,ALTO_PLATAFORMA,posX,posY,angulo),false){
 	this->largo = largo;
 	this->myVista = new VistaFigAgrandable(this);
 }
